remove duplicated enum
